Subject as primary data storage
To avoid having to hold all images in memory (would not scale well)
the dataset will hold a list of subjects, which are duplicated and
split into train, test, and validation sets. Images can still be
loaded as batches or in entiretey, but are now loaded in a totally
lazy way.